feat(NavigationBar.vue, config.ts): add switch to network link and name based on CHAIN_ID
Add a new link to the NavigationBar component that allows users to switch to a different network. The link's name and URL are determined by the CHAIN_ID value in the config file. If the CHAIN_ID is equal to HARBINGER_CHAIN_ID, the link's name is "Mainnet" and the URL is "https://koinosblocks.com". Otherwise, the link's name is "Harbinger" and the URL is "https://harbinger.koinosblocks.com".